fix: [NA] Folder management
