--- lib/auth/account/components/mycolors.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-
-class AppColors {
-  static const Color primary = Color(0xFF393939); 
-  static const Color secondary = Color(0xFFFFAC4B);
-
-  static const Color textColor = Color(0xFFFFFFFF); 
-  static const Color subText = Color(0xFFC1C1C1);
-}
