+++ lib/services/auth_gate.dart
