+++ lib/services/auth_service.dart
