+++ lib/theme/dark_mode.dart
@@ -1,13 +1,14 @@
 import 'package:flutter/material.dart';
 
 ThemeData darkMode = ThemeData(
-    brightness: Brightness.dark,
-    colorScheme: ColorScheme.dark(
-        surface: Colors.grey.shade900,
-        primary: Colors.grey.shade800,
-        secondary: Colors.grey.shade700,
-        inversePrimary: Colors.grey.shade300),
-    textTheme: ThemeData.dark().textTheme.apply(
-          bodyColor: Colors.grey[300],
-          displayColor: Colors.white,
-        ));
+  brightness: Brightness.dark,
+  colorScheme: ColorScheme.dark(
+      surface: Colors.grey.shade900,
+      primary: Colors.grey.shade800,
+      secondary: Colors.grey.shade700,
+      inversePrimary: Colors.grey.shade300),
+  textTheme: ThemeData.dark().textTheme.apply(
+        bodyColor: Colors.grey[300],
+        displayColor: Colors.white,
+      ),
+);
